move contributors format stuff to a templatetag
