+++ conventions/templatetags/display_filters.py
@@ -85,3 +85,31 @@ def display_gender_terminaison(convention: Convention) -> str:
     if convention.is_avenant():
         return ""
     return "e"
+
+
+@register.filter
+def display_contributors(convention: Convention) -> str:
+    result = ""
+
+    contributors = convention.get_contributors()
+
+    if not contributors["instructeurs"] and not contributors["bailleurs"]:
+        return "aucun contributeur connu pour l'instant"
+
+    if len(contributors["instructeurs"]) > 0:
+        result += "pour l'instruction "
+        for i, instructeur in enumerate(contributors["instructeurs"]):
+            result += f"{instructeur[0]} {instructeur[1]}"
+            if i != len(contributors["instructeurs"]) - 1:
+                result += ", "
+
+    if len(contributors["bailleurs"]) > 0:
+        if len(contributors["instructeurs"]) > 0:
+            result += " et "
+        result += "pour le bailleur "
+        for i, bailleur in enumerate(contributors["bailleurs"]):
+            result += f"{bailleur[0]} {bailleur[1]}"
+            if i != len(contributors["bailleurs"]) - 1:
+                result += ", "
+
+    return result
